prod - update parser + create graph img simplified
Parser only add terminals nodes to the graph, and not the non terminals
nodes. This is a simplification, and it is not a problem for the
parser. The graph is still correct.

Python programm added to create a graph image from a json file and
graphviz. Can be used to debug the parser.